chore(Refiner): move sdk loader to different file
