+++ src/integrations/Refiner/browser.js
@@ -1,7 +1,7 @@
 /* eslint-disable */
 import logger from '../../utils/logUtil';
-import { LOAD_ORIGIN } from '../../utils/ScriptLoader';
 import { NAME } from './constants';
+import { loader } from './loader';
 import { replaceUserTraits, replaceAccountTraits } from './utils';
 
 class Refiner {
@@ -24,15 +24,7 @@ class Refiner {
     this._refiner = function () {
       window._refinerQueue.push(arguments);
     };
-    (function () {
-      var a = document.createElement('script');
-      a.setAttribute('data-loader', LOAD_ORIGIN);
-      a.type = 'text/javascript';
-      a.async = !0;
-      a.src = 'https://js.refiner.io/v001/client.js';
-      var b = document.getElementsByTagName('script')[0];
-      b.parentNode.insertBefore(a, b);
-    })();
+    loader();
     this._refiner('setProject', this.apiKey);
   }
